Best error handling, data persistence and validation

+++ src/main/java/edu/test/poke_api/Dto/ProgressLoadDto.java
@@ -1,27 +1,27 @@
 package edu.test.poke_api.Dto;
 
 public class ProgressLoadDto {
-    int totalCount;
-    int dataLoaded;
-    int percentage;
+    long totalCount;
+    long dataLoaded;
+    long percentage;
 
     public ProgressLoadDto() {}
 
-    public ProgressLoadDto(int count, int currentLoad) {
+    public ProgressLoadDto(long count, long currentLoad) {
         totalCount = count;
         dataLoaded = currentLoad;
         percentage =  count > 0 ? currentLoad * 100 / count : 0;
     }
 
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
-    public int getDataLoaded() {
+    public long getDataLoaded() {
         return dataLoaded;
     }
 
-    public int getPercentage() {
+    public long getPercentage() {
         return percentage;
     }
