+++ src/main/java/edu/test/poke_api/models/PokeRecord.java
@@ -1,4 +1,4 @@
 package edu.test.poke_api.models;
 
-public record PokeRecord(int base_experience, int height, Integer id, boolean is_default, String name, int order, SpritesRecord sprites, int weight) {
+public record PokeRecord(int base_experience, int height, Long id, boolean is_default, String name, int order, SpritesRecord sprites, int weight) {
 }
